remove ppr
This change removes PPR. ATM when deployed to vercel PPR will produce fallback shells for routes wiht dynamic parameters. The issue page is a dynamic parameter page and so rather than upgrading each page to a specific route shell it just continuously serves the fallback shell. However in this app's setup there is no Suspense boundary nor dynamic content and the fallback shell is empty because it accesses an unknown param. We're discussing changes to the route shell upgrade heuristic for PPR so that we can better serve this style of application automatically while not aggressively building route shells for routes with high cardinality params in them that will rarely get reused. In the meantime I'm removing PPR b/c it isn't actively being utilized for any observable purpose.

By removing PPR the page will be produced entirely statically. Unknown pages will be blockign prerenders but after the initial prerender it will be fully static for the lifetime of the deployment